core/standards/css3/image-fit/js/inherited/006.html

<!doctype html>
<title>'inherit'ing from &lt;img></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  var e = document.createElement('img');
  e.appendChild(document.createElement('div'));
  e.firstChild.style.OObjectFit = 'inherit';
  e.firstChild.style.OObjectPosition = 'inherit';
  document.body.appendChild(e);
  assert_equals(getComputedStyle(e.firstChild, '').OObjectFit, 'auto');
  assert_equals(getComputedStyle(e.firstChild, '').OObjectPosition, '50% 50%');
});
</script>
